several changes and bugfixes...GEMS coupling seems to work, but transport does not proceed.....
git-svn-id: svn+ssh://savannah01.psi.ch/repos/gems/GEMSTANDALONE/trunk@149 9ff7d5ee-7a19-0410-ab54-ec97d5369f81

+++ mcotac-gem/setpar_.c
@@ -55,11 +55,11 @@
 #include "gwheader.h"
 
 #ifdef __PGI
-   void partid_( long npmax, int nbox, double xmin, double xmax,
+   void partid_( int npmax, int nbox, double xmin, double xmax,
 	                        int partib[51], double dx[51+2], double partx[50000])
 
 #else
-   void partid( long npmax, int nbox, double xmin, double xmax,
+   void partid( int npmax, int nbox, double xmin, double xmax,
 	                        int partib[51], double dx[51+2], double partx[50000])
 
 #endif
@@ -107,11 +107,11 @@
 #include "gwheader.h"
 
 #ifdef __PGI
- void concver_(long npmax,int nbox,double dx[51+2],double bn[51][10],
+ void concver_(int npmax,int nbox,double dx[51+2],double bn[51][10],
 			 double cn[51][25],int partib[51],double partx[50000],
 			 double partic[10+25][50000],int ismooth,int m1,int m2)
 #else
- void concver(long npmax,int nbox,double dx[51+2],double bn[51][10],
+ void concver(int npmax,int nbox,double dx[51+2],double bn[51][10],
 			 double cn[51][25],int partib[51],double partx[50000],
 			 double partic[10+25][50000],int ismooth,int m1,int m2)
 #endif
